backend/pyproject.toml
[tool.poetry]
name = "tutorrag"
version = "0.1.0"
description = "Grounded RAG backend for self-learning from personal study materials."
authors = ["OpenAI User <user@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pydantic = "^2.6.5"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.30"
aiosqlite = "^0.20.0"
python-dotenv = "^1.0.1"
sentence-transformers = "^2.6.0"
pypdf = "^4.2.0"
PyPDF2 = "^3.0.1"
python-docx = "^1.1.2"
python-pptx = "^0.6.23"
alembic = "^1.18.4"
rank-bm25 = "^0.2.2"
qdrant-client = "^1.17.1"
psycopg2-binary = "^2.9.11"
litellm = "^1.83.4"
pyjwt = "^2.12.1"
cryptography = "^46.0.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.7"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"